requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.0.0
openai==1.30.5
openai-whisper==20250625
python-dotenv==1.0.0
pydub==0.25.1
easyocr==1.7.0
scikit-learn==1.3.0
pandas==2.0.3
fastapi==0.115.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.10.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.24.1
streamlit==1.28.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.3.0
isort==5.12.0
flake8==6.1.0
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers==4.46.0
dashscope>=1.0.0
requests>=2.31.0
speechrecognition==3.10.0
pytesseract==0.3.10
apscheduler>=3.10.0
librosa>=0.10.0
soundfile>=0.12.0

# 单独处理 chromadb 和 sentence-transformers（手动安装）